### Changes

- Synchronize algorithms with 1.0.7
- Merge baseDependency, baseSubscriber into single signal struct, this will avoid type conversion in reactive_systems.
- Use flags directly to determine the type of dep.
- Instead of using the interface as a reference to dep and sub, the interface is exposed via signal.ref.

### Benchmark

main branch

```
+------------------------+-------------+------------+-------------+-------------+-------------+
| BENCHMARK              |         AVG |        MIN |         P75 |         P99 |         MAX |
+------------------------+-------------+------------+-------------+-------------+-------------+
| propagate: 1 * 1       |       332ns |      125ns |       167ns |       334ns |    16.542µs |
| propagate: 1 * 10      |       952ns |      750ns |         1µs |     1.292µs |         2µs |
| propagate: 1 * 100     |     5.947µs |    5.833µs |     5.958µs |     6.417µs |      6.75µs |
| propagate: 1 * 1000    |    54.862µs |   53.042µs |    54.791µs |    59.083µs |    68.125µs |
| propagate: 10 * 1      |     1.213µs |    1.166µs |     1.209µs |     1.292µs |     1.708µs |
| propagate: 10 * 10     |     6.576µs |    6.416µs |     6.791µs |     7.208µs |     7.209µs |
| propagate: 10 * 100    |    60.499µs |       58µs |     59.75µs |    82.833µs |    83.208µs |
| propagate: 10 * 1000   |   551.053µs |  536.458µs |   550.084µs |   636.417µs |   638.375µs |
| propagate: 100 * 1     |     13.67µs |   11.667µs |    12.209µs |    51.083µs |   121.834µs |
| propagate: 100 * 10    |    72.248µs |   69.208µs |    72.584µs |    76.125µs |    92.458µs |
| propagate: 100 * 100   |   590.335µs |  581.667µs |   590.125µs |   618.125µs |   644.417µs |
| propagate: 100 * 1000  |   5.48764ms | 5.456917ms |  5.488709ms |   5.74375ms |  5.937416ms |
| propagate: 1000 * 1    |   117.818µs |  116.958µs |     117.5µs |   128.458µs |   134.583µs |
| propagate: 1000 * 10   |   701.335µs |  693.417µs |   703.208µs |   714.708µs |   802.542µs |
| propagate: 1000 * 100  |  6.001302ms | 5.936709ms |  5.986375ms |  6.330875ms |  7.378917ms |
| propagate: 1000 * 1000 | 56.932538ms | 56.52575ms | 56.950541ms | 58.889416ms | 59.726667ms |
+------------------------+-------------+------------+-------------+-------------+-------------+
```

This PR

```
+------------------------+-------------+-------------+-------------+-------------+-------------+
| BENCHMARK              |         AVG |         MIN |         P75 |         P99 |         MAX |
+------------------------+-------------+-------------+-------------+-------------+-------------+
| propagate: 1 * 1       |       361ns |       125ns |       167ns |       708ns |    19.542µs |
| propagate: 1 * 10      |       560ns |       458ns |       625ns |     1.083µs |     1.084µs |
| propagate: 1 * 100     |     4.122µs |     3.875µs |     3.959µs |     4.792µs |    20.542µs |
| propagate: 1 * 1000    |     46.66µs |    37.291µs |    56.584µs |     59.25µs |    83.583µs |
| propagate: 10 * 1      |     1.013µs |       916ns |         1µs |     1.583µs |     2.209µs |
| propagate: 10 * 10     |     4.303µs |     4.167µs |     4.333µs |     4.792µs |     4.834µs |
| propagate: 10 * 100    |    49.144µs |    39.917µs |    50.375µs |    54.083µs |   169.792µs |
| propagate: 10 * 1000   |   292.419µs |   269.459µs |   311.667µs |   383.917µs |   901.709µs |
| propagate: 100 * 1     |     5.846µs |     5.417µs |     5.708µs |    11.291µs |    11.959µs |
| propagate: 100 * 10    |    34.407µs |    32.458µs |    34.584µs |    41.459µs |    58.708µs |
| propagate: 100 * 100   |   300.712µs |   296.542µs |   301.417µs |   313.459µs |   330.083µs |
| propagate: 100 * 1000  |  2.803827ms |  2.755417ms |   2.79875ms |  3.175875ms |   3.21925ms |
| propagate: 1000 * 1    |    65.128µs |     58.25µs |    62.166µs |    88.292µs |   439.625µs |
| propagate: 1000 * 10   |   339.583µs |   324.459µs |     333.5µs |   398.083µs |  1.011583ms |
| propagate: 1000 * 100  |  3.141568ms |  3.085792ms |  3.134625ms |  3.653125ms |  3.900417ms |
| propagate: 1000 * 1000 | 30.555207ms | 30.248041ms | 30.541458ms | 32.487333ms | 32.699667ms |
+------------------------+-------------+-------------+-------------+-------------+-------------+
```